Add NavMesh asset for Jobs Scene, update URP settings, and improve job structures
Integrated NavMesh configuration for "Jobs Scene." Updated URP settings to include enhanced reflection probe and shadow-related configurations. Optimized `MultiplyJob` and `MapCreationJob` for better performance and maintainability, including struct updates and coroutine-based execution.

+++ Assets/Scripts/Jobs/MapCreationJob.cs
@@ -2,7 +2,6 @@
 using Unity.Collections;
 using Unity.Jobs;
 
-
 [BurstCompile]
 public struct MapCreationJob : IJobParallelFor
 {
@@ -12,18 +11,13 @@ public struct MapCreationJob : IJobParallelFor
     public NativeArray<float> resourcesResultArray;
     public NativeArray<float> terrainResultArray;
     
-    // public void Execute(int index)
-    // {
-    //     itemsResultArray[index] = rand.NextInt(1, 10);
-    // }
-    
     public void Execute(int index)
     {
         Unity.Mathematics.Random threadRandom = new Unity.Mathematics.Random(15 + (uint)index);
-        itemsResultArray[index] = threadRandom.NextInt(1, 10);
-        npcResultArray[index] = threadRandom.NextInt(1, 10);
-        strucutresResultArray[index] = threadRandom.NextInt(1, 10);
-        resourcesResultArray[index] = threadRandom.NextInt(1, 10);
-        terrainResultArray[index] = threadRandom.NextInt(1, 10);
+        itemsResultArray[index] = threadRandom.NextFloat(1, 10);
+        npcResultArray[index] = threadRandom.NextFloat(1, 10);
+        strucutresResultArray[index] = threadRandom.NextFloat(1, 10);
+        resourcesResultArray[index] = threadRandom.NextFloat(1, 10);
+        terrainResultArray[index] = threadRandom.NextFloat(1, 10);
     }
 }
